requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
asttokens==3.0.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
deepdiff==8.0.1
docker-pycreds==0.4.0
docopt-ng==0.9.0
enum34==1.1.10
exceptiongroup==1.2.2
executing==2.1.0
fancycompleter==0.9.1
Farama-Notifications==0.0.4
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.10.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.0.0
idna==3.10
imageio==2.36.1
iniconfig==2.0.0
ipdb==0.13.13
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonpickle==4.0.0
kiwisolver==1.4.7
lbforaging==2.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mock==5.1.0
mpmath==1.3.0
mpyq==0.2.5
munch==4.0.0
networkx==3.4.2
numpy==2.2.0
orderly-set==5.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pbr==6.1.0
pdbpp==0.10.3
pettingzoo==1.24.3
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
portpicker==1.6.0
probscale==0.2.5
prompt_toolkit==3.0.48
protobuf==3.20.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
py-cpuinfo==9.0.0
pydantic==2.10.3
pydantic_core==2.27.1
pygame==2.6.1
pyglet==1.5.27
Pygments==2.18.0
pyparsing==3.2.0
pyrepl==0.9.0
PySC2==3.0.0
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==5.3.1
requests==2.32.3
Rtree==1.0.0
rware==2.0.0
s2clientprotocol==4.10.1.75800.0
s2protocol==5.0.14.93333.0
sacred==0.8.7
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
sentry-sdk==2.19.2
setproctitle==1.3.4
Shimmy==2.0.0
six==1.17.0
sk-video==1.1.10
smmap==5.0.1
snakeviz==2.2.2
stack-data==0.6.3
sympy==1.13.1
tensorboard-logger==0.1.0
threadpoolctl==3.5.0
tomli==2.2.1
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.19.1
wcwidth==0.2.13
websocket-client==1.8.0
wrapt==1.17.0